requirements.txt
# 核心依赖
mcp>=0.2.5
python-dotenv>=1.0.0
structlog>=23.1.0
pydantic>=2.0.0
PyYAML>=6.0.1

# 图像处理依赖
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0

# 网络和并发依赖
aiohttp>=3.8.0
aiofiles>=23.2.0
asyncio>=3.4.3

# 数据处理依赖
pandas>=2.0.0
jsonschema>=4.0.0

# 缓存依赖
cachetools>=5.3.0

# OCR引擎依赖 - 推荐使用RapidOCR
rapidocr-onnxruntime>=1.4.4  # 推荐：轻量级、高性能OCR引擎（最新版本）
onnxruntime>=1.15.0

# ModelScope依赖（可选，仅在使用ModelScope引擎时需要）
# modelscope>=1.28.0
# addict>=2.4.0
# datasets>=2.20.0,<3.0.0 